pyproject.toml
[tool.poetry]
name = "my_github"
version = "0.1.0"
description = ""
authors = ["ispraf <ispraf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
my-github = "my_github.main:my_github"